requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
limits==2.6.2
packaging==21.3
pydantic==1.9.0
pylogbeat==2.0.0
pyparsing==3.0.9
python-dotenv==0.19.2
python-logstash-async==2.5.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
wrapt==1.14.1
